Module Flattened, README.rst added. :Docstring functions

# Conflicts:
#	patella/__pycache__/htmlparser.cpython-36.pyc

+++ patella/htmlparser.py
@@ -1,78 +1,63 @@
+# -*- coding: utf-8 -*-
 #!/usr/bin/env python
-"""Html parsing and dataframe utility"""
+
+"""
+This module has two important functions:
+
+htmlparser.find_download_links(): this function is used to scrape webpages for download links and html tables,
+                                  allowing users to find and download items from a page.
+
+htmlparser.compare(): this function is used to compare some data frame with a date column in it to the years in office
+                      of the relevant president or party in office at the time.
+"""
+
+'''\/ Local Imports \/'''
+from patella.tablereader import get_fe
 
 ''' \/ Third-Party Imports \/ '''
-from bokeh.plotting import figure, show
+from bokeh.plotting import figure
 import bokeh.palettes as palettes
 from bokeh.embed import components
 from flask import render_template
 from urllib.parse import urlparse
 from bs4 import BeautifulSoup
 import pandas as pd
-import os, sys
+import os
 import urllib.request
 import logging
-import matplotlib.pyplot as plt
-plt.style.use('ggplot')
 import datetime
 from dateparser import parse
 
-
-'''Some useful data frame commands'''
-
-logging.basicConfig(filename='webservice.log', level=logging.DEBUG ) # create a logging/print function for me
-fe_url = 'http://www.enchantedlearning.com/history/us/pres/list.shtml'
-fe_table = pd.read_html(fe_url, match='Vice-President', flavor='bs4', header=0, index_col=2, parse_dates=True)[0]
-fe_names = []
-index_list = fe_table.index.get_level_values(0).values.tolist()
-year_list = []
+logging.basicConfig(filename='webservice.log', level=logging.DEBUG )  # set up the logger at level debug
 
 
 def prlog(msg, log=True, prnt=False, level='ERROR'):
     """
+    :type log: bool
     :param msg: msg to print/log
     :param log: if the function should log the msg or not
     :param prnt: if the function should print the msg or not
     :param level: what level the logger records the message at
     :return: None
     """
-    if prnt == True:
+
+    if prnt:
         print(msg)
-    elif log == True:
+    elif log:
+        print("logging." + level)
         logging.error(msg)
 
 
-for item in index_list:
-    term_num = item.split('-')[0]
-    year_list.append(term_num)
-
-fe_table['Year'] = year_list
-fe_table.set_index('President', drop=False, inplace=True)
-
-for idx, item in enumerate(fe_table.index.get_level_values(0).values.tolist()):
-    fe_name = ''
-    for char in item:
-        if char.isalpha() or char == ' ':
-            fe_name = fe_name + char
-    fe_names.append(fe_name)
-
-fe_table['President'] = fe_names
-fe_table.set_index('Year', drop=True, inplace=True)
-
-
-def get_fe():
-    return fe_table
-
-
-def download_approved(ext, filep):  # A function that returns true if the file extension is correct AND the file is readable
+# A function that returns true if the file extension is correct AND the file is readable
+def download_approved(ext, filep):
     """
     :param ext: the expected file extension of the file
     :param filep: the filepath of the file
     :return: True if approved, False if not
     """
     approved_exts = ['.csv', '.tsv']
-    delim = ''
-    problem = 'none'
+    delim = ''  # namespace scope expansion
+    problem = 'none'  # by default, no problems
     for item in approved_exts:
         if ext == item:
             if item is '.csv':
@@ -80,17 +65,22 @@ def download_approved(ext, filep):  # A function that returns true if the file e
             elif item is '.tsv':
                 delim = '\t'
             else:
-                problem = 'some'
+                problem = 'some'  # if problem, some problem
     try:
         filep.encode('utf-8').strip()
         pd.read_table(filep, delim, header=0, engine='python')
-    except Exception as exc:
+    except Exception as exc:  # if problem, some problem
         logging.error(exc)
         problem = 'some'
     if problem == 'none':
         return True
     elif problem == 'some':
         return False
+
+
+# a function which defines the previous elemenet in a list - makes the code more readable\
+def previous_element(_list, index):
+    return _list[index-1]
 
 
 # A function which looks around webpages for html tables and .csv/.tsv download links and downloads them.
@@ -102,15 +92,14 @@ def find_download_links(url, filetype, output_name, in_number=0, download=False,
     :param output_name: name of the file to be downloaded
     :param in_number: the number by order of the link in the url
     :param download: if the function should download a file or just return
-    :param render: if the template should be rendered or the function should just return
     :param clobber: if files with the same name should be clobbered
-    :param urlth: the path the url takes to render
-    :return: a dictionary that returns an error as 'error', the outputname as 'filename' and the list of links in the url as 'href_list'
+    :return: a dictionary that returns an error as 'error', the outputname as 'filename'
+             and the list of links in the url as 'href_list'
     """
     error = 'None'
     p_url = urlparse(url)
-    domain = '{urm.scheme}://{urm.netloc}'.format(urm=p_url)
-    dl_name = output_name
+    domain = '{urm.scheme}://{urm.netloc}'.format(urm=p_url)  # using outdated urllib1 - this line obtains domain
+    dl_name = output_name  # set to the output name by default if no transformations are needed
     link_list = []
     no_tags = ''
     ext_length = len(filetype)
@@ -122,7 +111,7 @@ def find_download_links(url, filetype, output_name, in_number=0, download=False,
     # First and foremost, check if the user is entering a valid data format
     if extension == filetype:
         link_list.append(url)   # Before anything, check if the url entered IS a dl link
-        if download:  # check if directory exists, check if file is parseable, check allowed file extensions
+        if download and clobber:  # check if directory exists, check if file is parseable, check allowed file extensions
             urllib.request.urlretrieve(url, output_path)
             if not download_approved(extension, output_path):
                 os.remove(output_path)
@@ -138,8 +127,8 @@ def find_download_links(url, filetype, output_name, in_number=0, download=False,
                 if 'http://' not in no_tags and 'https://' not in no_tags:   # If no first part of the url, add it
                     no_tags = domain + '/' + no_tags
                 link_list.append(no_tags)
-        if download:
-            if no_tags != None:   # Null check
+        if download and clobber:
+            if no_tags is not None:   # Null check
                 if len(link_list) == 1:
                     prlog("One link found: " + no_tags)
                     urllib.request.urlretrieve(no_tags, output_path)
@@ -164,13 +153,15 @@ def find_download_links(url, filetype, output_name, in_number=0, download=False,
             elif filetype not in str(no_tags):
                 error = 'No file found for that extension (' + filetype + ')'
     prlog(error)
-    return { 'error':error, 'download_name':dl_name, 'href_list':link_list }
+    return {'error': error, 'download_name': dl_name, 'href_list': link_list}
 
 
-def file_to_htmltable(filepath):
-        dataframe = pd.read_table(filepath, ',', header=0, engine='python')
-        htmltable = dataframe.to_html(bold_rows=True, escape=True)
-        return htmltable
+# simple function to take a file from a file path and generate an html table
+def file_to_htmltable(filepath, delim=','):
+    dataframe = pd.read_table(filepath, delim, header=0, engine='python')
+    htmltable = dataframe.to_html(bold_rows=True, escape=True)
+    return htmltable
+
 
 # A function that takes a dataframe for input and compares it to the change by year as a political party held power
 def compare(df1, col, title, x_lb, y_lb, fedf=get_fe(), year_col='Year', html='plotlocal.html', render=True, urlth='patella'):
@@ -181,36 +172,45 @@ def compare(df1, col, title, x_lb, y_lb, fedf=get_fe(), year_col='Year', html='p
     :param title: the title of the plot to be exported.
     :param x_lb: the title of the x axis for the exported plot.
     :param y_lb: the title of the y axis for the exported plot.
-    :param html: the html page to be rendered.
+    :param year_col: the titel of column of the data frame input to set as the index
+    :param html: the html page to be rendered
+    :param render: if the function should return js components for a bokeh plot or just the input/resulting dataframes
+    :param urlth: the path the url takes to render
     :return: a render_template object which sends a bokeh js/html/css plot to the page as well as a special div.
     """
 
-    #### V Some Global variables V  ####
-    data_col = int(col)
-    df1.set_index(year_col, drop=True, inplace=True)
+    ''' V Some Global variables V  '''
+    data_col = int(col)-1  # subtract 1 from it so the column counting starts from one in the input
+    df1.set_index(year_col, drop=True, inplace=True) # if year_col and type(year_col) is "string" else prlog('year col not string')
     ind_list = [parse(str(int(x))).year for x in df1.index.get_level_values(0).values.tolist()]
     office_yr_list = []
     party_office_list = []
-    year_num = 1
-    year_num_of = 1
+    yrs_in_office = 1
+    party_yrs_in_office = 1
     term_st_list = fedf.index.get_level_values(0)
     totalfe_list = []
     totalparty_list = []
 
-    # populate a list with tuples of the year and the index where the year changes
-    year_and_interval_list = [(i, ind_list[i-1]) for i, year in enumerate(ind_list) if i >=1 and ind_list[i-1] != ind_list[i]]
+    # populate a list with tuples of the year and the index where the year changes by looping through the index of the
+    # data frame, which is the year column. This spits out a tuple that looks like (years, number of data points)
+    # afterwards , the lists interval and years_list are used to store the parts of the tuple individually
+    year_and_interval_list = [(i, previous_element(ind_list, i)) for i, year in
+                              enumerate(ind_list) if i >= 1 and previous_element(ind_list, i) != ind_list[i]]
     interval = [tpl[0] for tpl in year_and_interval_list]
     years_list = [tpl[1] for tpl in year_and_interval_list]
     total_chg = [(df1.iloc[0, data_col]-df1.iloc[interval[i], data_col]) if i == 0
-                 else (df1.iloc[interval[i-1], data_col] - df1.iloc[interval[i], data_col]) if interval[i] < len(ind_list)
-                 else (df1.iloc[interval[i-1], data_col] - df1.iloc[(len(ind_list) - interval[i-1]), data_col])
+                 else (df1.iloc[previous_element(interval, i), data_col] - df1.iloc[interval[i], data_col]) if interval[i] < len(ind_list)
+                 else (df1.iloc[previous_element(interval, i), data_col] - df1.iloc[(len(ind_list) - previous_element(interval, i)), data_col])
                  for i, item in enumerate(years_list)]
-    # A list of variation from year to year
-    cov_list = [total_chg[i]/total_chg[i-1]*100 if total_chg[i-1] != 0 else 0 for i, n in enumerate(years_list)]
 
-    # Create lists of all the parties and years over all the us years
+    # A list of variation from year to year, created by dividing the list at index by the previous element,
+    # then multiplying by 100 to get a percentage value.
+    cov_list = [total_chg[i]/previous_element(total_chg, i)*100 if previous_element(total_chg, i) != 0
+                else 0 for i, n in enumerate(years_list)]
+
+    # Create lists of all the parties and years over all the us years by finding the distance from from the first
+    # presidential term served to the last and appending incremented years in that range to the totalfe totalparty lists
     for idx, item in enumerate(term_st_list):
-        dist = 0
         if idx != len(term_st_list)-1:
             dist = int(term_st_list[idx+1]) - int(item)
         else:
@@ -220,24 +220,25 @@ def compare(df1, col, title, x_lb, y_lb, fedf=get_fe(), year_col='Year', html='p
             totalparty_list.append(fedf.iloc[idx, 1])
     # A list of all years there have been presidents
     us_years = [yr+int(term_st_list[0]) for yr in range(int(term_st_list[-1]) - int(term_st_list[0]))]
-    # A list of all the parties over every year of us presidency
+    # A list of all the parties over every year of the data set
     party_list = [totalparty_list[i] for i, item in enumerate(us_years) for years in years_list if years == item]
-    # A list of all the fes over every year of us presidency
+    # A list of all the fes over every year of the data set
     fe_list = [totalfe_list[i] for i, item in enumerate(us_years) for years in years_list if years == item]
-    # Make the 'years in office' and 'party in office' lists
-    # list  comprehension possible?
+    # Make the 'years in office' and 'party in office' lists by finding the number of consecutive
+    # a party/fe was in office
+    # list comprehension possible?
     for idx, item in enumerate(fe_list):
         if idx >= 1 and fe_list[idx - 1] == fe_list[idx]:   # After at least one loop (to avoid oob error) and if the index doesnt change
-            year_num += 1
+            yrs_in_office += 1  # increment 1
         else:
-            year_num = 1
-        office_yr_list.append(year_num)   # append the year number
+            yrs_in_office = 1  # reset to year 1
+        office_yr_list.append(yrs_in_office)   # append the year number
     for idx, item in enumerate(party_list):
         if idx >= 1 and party_list[idx - 1] == party_list[idx]:  # After at least one loop (to avoid oob error) and if the index doesnt change
-            year_num_of += 1
+            party_yrs_in_office += 1  # increment 1
         else:
-            year_num_of = 1
-        party_office_list.append(year_num_of)   # append the year number
+            party_yrs_in_office = 1  # reset to year 1
+        party_office_list.append(party_yrs_in_office)   # append the year number
 
     # Create the dataframe 'plotframe' that will be used to create the bokeh graph
     plotframe = pd.DataFrame({'foo': []})
@@ -248,25 +249,24 @@ def compare(df1, col, title, x_lb, y_lb, fedf=get_fe(), year_col='Year', html='p
     plotframe['First Executive'] = fe_list
     plotframe['Years in Office'] = office_yr_list
     plotframe['Years Party in Office'] = party_office_list
-    plotframe.set_index('foo', drop=True, inplace=True)
-    plotframe.set_index('Years', drop=False, inplace=True)
+    plotframe.set_index('foo', drop=True, inplace=True)  # remove stock column in a roundabout way
+    plotframe.set_index('Years', drop=False, inplace=True)  # remove the year column and set as index
     grouped = plotframe.groupby('First Executive')
 
-    #''' Bokeh Plot ''' ''' -- being converted to Plotly Javascript --
+    # ''' Bokeh Plot ''' ''' -- soon to be converted to Plotly Javascript --
     p = figure(title=title, x_axis_label=x_lb, y_axis_label=y_lb, tools="pan,box_zoom,reset,save", toolbar_location='below', toolbar_sticky=False)
-    palette = [palettes.viridis(len(grouped))[i] for i, _ in enumerate(grouped)]  # use the plasma palette
-    imdex = -1
-    for name, data in grouped:
-        imdex += 1
-        p.line(x=data['Years in Office'], y=data['Percent Change'], color=palette[imdex], legend=name) #palette[imdex]
-    script, div = components(p) # split the graph into JSON/JS components script and div
-    #'''            ''' ''' -- being converted to Plotly Javascript --
+    # use the plasma palette from bokeh using the number of lines in the figure to create a palette from plasma256
+    palette = [palettes.viridis(len(grouped))[i] for i, _ in enumerate(grouped)]
+    for idx, (name, data) in enumerate(grouped):
+        p.line(x=data['Years in Office'], y=data['Percent Change'], color=palette[idx], legend=name)
+    script, div = components(p)  # split the graph into JSON/JS components script and div
+    # '''            ''' ''' -- soon to be converted to Plotly Javascript --
 
-
-
-
-    df_htmltable = plotframe.to_html(bold_rows=True, escape=True, classes='dftable') # convert the datframe to html for easy display\
+    # convert the datframe to html for easy display\
+    df_htmltable = plotframe.to_html(bold_rows=True, escape=True, classes='dftable')
     if render:
-        return render_template(html, script=script, div=div, table=df_htmltable, var=urlth)  # If rendering, display the web page
+        # If rendering, display the web page
+        return render_template(html, script=script, div=div, table=df_htmltable, var=urlth)
     else:
-        return {'plotframe':plotframe, 'input_frame':df1}  # otherwise, return the dataframe and the input dataframe
+        # otherwise, return the result dataframe and the input dataframe
+        return {'plotframe': plotframe, 'input_frame': df1}
